Output differences between generated and reference files in a failed test.
One difficulty of running the tests is that we don't see the effect
source of the differences between files. The new command line does output
them directly so that diagnosing a failure is faster and more convenient.

Moreover, the command to compare the results has been improved to avoid
bogus differences due to empty lines, spacing, or simply CR in windows vs.
unix files.