Removed *-e files and bug that generated them and modified references.
A missing argument in sed -i -e made sed use -e as the extension for the
backupfile. Moreover, the in-place change that test_line wanted to do using sed
was intended to be previously applied by hand in all the reference files.
Unfortunately that step was forgotten. But now it has been done.