Improved the graphical output, now in pdf format.
The graphical output in POY was good, but not good enough. The edges in the
trees where completely disconnected, small dots where added to fill gaps in the
corners of the tree, the strings had some times a separated character at the
end, the margins where different on each page, and the postscript format was not
good for windows users who had to struggle with ghostview for a quick look.
Moreover, the current output (PS), does not support cross-references, something
that we need to produce a reasonable diagnosis (which is the ultimate reason
that triggered this comprehensive solution).

All these problems have been addressed in this commit by:
0. Incorporate camlpdf-0.3 to the code base.
1. Replace the ps with a pdf output.
2. Make the branches of the tree paths connecting the upper and lower children,
while all the intermediate children are horizontal lines ending in the vertical
portion of the path.
3. Cleanup the output by using the native pdf strings, avoiding the breakage of
the terminal names (this was an effect of Graphps).

The LICENSE file has been updated accordingly, and the documentation too, to
reflect the new format generated by POY.