Fixed configuration and compilation under windows using ocamlbuild.
The compilation in windows using ocamlbuild was broken. It now requires OCaml
version 3.11.0, and the rules have been broken in link, and compile flags, so
that different tools can be used at each stage (which is a requirement for the
ocaml windows port, that uses flexdll in the linking step).

use