Updating Likelihood scripts
There was a critical issue in the previous design. We could not determine a way
to test the output of the scripts. It was determined that FILENAME* should be
used. This makes sense but causes issues since these are used in the scripts,
but can contain characters that are not allowed in filenames. So, I've changed
it so this is allowed. This involved moving a lot of the scripts around.