fixed issue with test-line when linebreaks happen in output
since fixing the format output, text properly wraps around on 80char margins.
This works great, but the comparison application in the test suite removes lines
with time differences, but if the time is too long (on slow machines), then the
line will break and that is the line removed, but in reality both need to be
removed for a proper comparison.